+++ tests/greedysearch_test.cpp
@@ -3,7 +3,7 @@
 #include "../include/vamana.h"
 
 
-Node* create_node(unsigned int id, const vector<double>& coords) {
+Node* create_node(unsigned int id, const vector<float>& coords) {
     Node* node = new Node();
     node->id = id;
     node->coords = coords;
